--- mugen-common/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "mugen-common"
-version = "0.1.0"
-authors = ["koopa1338 <sinner1991@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-serde = { version = "1.0.118", features = ["derive"] }
-diesel = { version = "1.4.5", features = ["postgres", "extras"] }
-dotenv = "0.15.0"
-chrono = { version = "0.4.19", features = ["serde"] }
