--- mugen-common/src/handler/mod.rs
@@ -1,2 +1,0 @@
-mod auth;
-mod register;
