+++ mugen-server/Cargo.toml
@@ -6,11 +6,17 @@ edition = "2018"
 
 [dependencies]
 actix-files = "0.5.0"
+actix-rt = "2.2.0"
+actix-service = "2.0.0"
 actix-web = "3.3.2"
-actix-rt = "1.1.1"
-actix-service = "1.0.6"
-env_logger = "0.8.2"
+actix-web-grants = "2.0.1"
+chrono = { version = "0.4.19", features = ["serde"] }
 clap = "2.33.3"
-log = "0.4.11"
-mugen-common = { path = "../mugen-common" }
-tokio = "1.0.1"
+diesel = { version = "1.4.6", features = ["postgres", "extras"] }
+dotenv = "0.15.0"
+env_logger = "0.8.3"
+log = "0.4.14"
+r2d2 = "0.8.9"
+r2d2_postgres = "0.18.0"
+serde = { version = "1.0.126", features = ["derive"] }
+tokio = "1.5.0"
