+++ mugen-server/src/main.rs
@@ -2,8 +2,9 @@
 extern crate log;
 
 use actix_files as fs;
-use actix_service::Service;
-use actix_web::{http, web, App, HttpResponse, HttpServer, Result as WebResult};
+use actix_web::{dev::Service, http, web, App, HttpResponse, HttpServer, Result as WebResult};
+use r2d2::Pool;
+use r2d2_postgres::{postgres::NoTls, PostgresConnectionManager};
 
 use clap::Arg;
 
@@ -36,8 +37,8 @@ fn get_config() -> AppConfig {
                 .takes_value(true),
         )
         .arg(
-            Arg::with_name("pages")
-                .long("pages")
+            Arg::with_name("static")
+                .long("static")
                 .value_name("DIR")
                 .required(true)
                 .help("Path to the directory containing HTML & static files")
@@ -76,10 +77,21 @@ async fn app_page(state: web::Data<AppState>) -> WebResult<fs::NamedFile> {
     Ok(fs::NamedFile::open(state.config.pages.join("index.html"))?)
 }
 
+pub fn get_db_pool() -> Result<Pool<PostgresConnectionManager<NoTls>>, r2d2::Error> {
+    let manager =
+        PostgresConnectionManager::new("host=localhost user=postgres".parse().unwrap(), NoTls);
+
+    r2d2::Pool::builder().max_size(15).build(manager)
+}
+
 #[actix_rt::main]
 async fn main() -> std::io::Result<()> {
     env_logger::init();
     let config = get_config();
+
+    let pool = get_db_pool().unwrap();
+
+    let _client = pool.get().unwrap();
 
     let addr = format!("127.0.0.1:{}", config.port);
     info!("Listening on {}", addr);
