+++ mugen-server/src/persistence/models/user.rs
@@ -23,22 +23,3 @@ pub struct User {
 }
 
 
-#[derive(Debug, Clone, Queryable)]
-pub struct DocumentQuery {
-    pub id: i64,
-    pub created: NaiveDateTime,
-    pub last_updated: Option<NaiveDateTime>,
-    pub filetype: String, // crate for filetype?
-    pub version: i32,
-    pub size: i64,
-}
-
-#[derive(Debug, Clone, Insertable)]
-#[table_name="documents"]
-pub struct Document {
-    pub created: NaiveDateTime,
-    pub last_updated: Option<NaiveDateTime>,
-    pub filetype: String, // crate for filetype?
-    pub version: i32,
-    pub size: i64,
-}
