+++ mugen-server/src/persistence/utils.rs
