+++ mugen-ui/Cargo.toml
@@ -8,8 +8,7 @@ edition = "2018"
 crate-type = ["cdylib"]
 
 [dependencies]
-# mugen-common = { path = "../mugen-common" }
-wasm-bindgen = "0.2.69"
-web-sys = { version = "0.3.46", features = ["HtmlElement", "HtmlInputElement", "Node"] }
+wasm-bindgen = "0.2.74"
+web-sys = { version = "0.3.51", features = ["HtmlElement", "HtmlInputElement", "Node"] }
 yew = { version = "0.17.4", features = ["web_sys"] }
-yew-router = { version = "0.14", features = ["web_sys"] }
+yew-router = { version = "0.14.0", features = ["web_sys"] }
