+++ static/index.html
@@ -3,15 +3,15 @@
   <head>
     <meta charset="utf-8" />
     <title>Mugen DMS</title>
-    <link rel="stylesheet" href="/static/css/uikit.min.css" />
-    <link rel="stylesheet" href="/static/css/mugen.css" />
+    <link rel="stylesheet" href="/css/uikit.min.css" />
+    <link rel="stylesheet" href="/css/mugen.css" />
     <script src="https://code.jquery.com/jquery-3.6.0.min.js" integrity="sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4=" crossorigin="anonymous"></script>
-    <script src="/static/js/uikit.min.js"></script>
-    <script src="/static/js/uikit-icons.min.js"></script>
+    <script src="/js/uikit.min.js"></script>
+    <script src="/js/uikit-icons.min.js"></script>
   </head>
   <body>
     <div id="app-container"></div>
-    <script type="module" src="/static/js/mugen.js"></script>
+    <script type="module" src="/js/mugen.js"></script>
   </body>
 </html>
